Set up new evaluation schme.
The ManualEvaluation class provides a framework for aggregating a set of ExtractionAnnotations
for a given test document set. Each time a different Multir Model is run the user should tag the
diff and merge with the previous annotations in order to calculate precision and yield.
Another minor change was I removed NIL links from NELArgumentIdentification.